Panel directly observes dataset displayed attr
Fixes #118.
Before this commit, the LayerPanel would send an event to the MapPanel when a
dataset had its displayed attribute changed.  Now the MapPanel listens directly
on the dataset collection, observing the displayed attribute of individual
datasets.  This also makes the UI more flexible by having the MapPanel no
longer depend on the LayerPanel.